chore!: replace old ContentTypeBadge with new
