--- packages/ndla-ui/src/ContentTypeBadge/index.ts
@@ -1,21 +1,0 @@
-/**
- * Copyright (c) 2018-present, NDLA.
- *
- * This source code is licensed under the GPLv3 license found in the
- * LICENSE file in the root directory of this source tree.
- *
- */
-
-import { ContentTypeBadge } from "./ContentTypeBadge";
-
-export {
-  SubjectMaterialBadge,
-  TasksAndActivitiesBadge,
-  AssessmentResourcesBadge,
-  SubjectBadge,
-  SourceMaterialBadge,
-  LearningPathBadge,
-  ConceptBadge,
-} from "./ContentTypeBadge";
-
-export default ContentTypeBadge;
