+++ packages/ndla-ui/src/ContentTypeBlockQuote/ContentTypeBlockQuote.tsx
@@ -8,7 +8,7 @@
 
 import { forwardRef } from "react";
 import { BlockQuote, BlockQuoteProps, BlockQuoteVariantProps } from "@ndla/primitives";
-import { ContentType } from "../ContentTypeBadge/ContentTypeBadgeNew";
+import { ContentType } from "../ContentTypeBadge/ContentTypeBadge";
 
 export type ContentTypeBlockQuoteVariant = "colored" | "neutral";
