+++ packages/ndla-ui/src/ContentTypeHero/ContentTypeHero.tsx
@@ -8,7 +8,7 @@
 
 import { forwardRef } from "react";
 import { Hero, HeroProps, HeroVariant } from "@ndla/primitives";
-import { ContentType } from "../ContentTypeBadge/ContentTypeBadgeNew";
+import { ContentType } from "../ContentTypeBadge/ContentTypeBadge";
 import * as contentTypes from "../model/ContentType";
 
 export const contentTypeToHeroMap: Record<ContentType, HeroVariant> = {
